Shrink and Continuous Compress
Added a function for shrinking the image by removing column and rows
which is a poor way to do it but it’s my final exam week and needed the
function ASAP so here it is, maybe I’ll improve that.

Also added a small loop and def’d compression as a function so sets of
images can be comprassed otomatically, just name it as 1, 2, 3 … etc.